Moving the Graphics*Object types to the `TerraFX.Graphics.Advanced` namespace
