+++ sources/Graphics/Core/Graphics/Advanced/GraphicsPipelineObject.cs
@@ -3,27 +3,29 @@
 using System;
 using static TerraFX.Utilities.ExceptionUtilities;
 
-namespace TerraFX.Graphics;
+namespace TerraFX.Graphics.Advanced;
 
-/// <summary>An object which is created for a graphics render pass.</summary>
-public abstract class GraphicsRenderPassObject
+/// <summary>An object which is created for a graphics pipeline.</summary>
+public abstract class GraphicsPipelineObject
 {
     private readonly GraphicsAdapter _adapter;
     private readonly GraphicsDevice _device;
+    private readonly GraphicsPipeline _pipeline;
     private readonly GraphicsRenderPass _renderPass;
     private readonly GraphicsService _service;
 
-    /// <summary>Initializes a new instance of the <see cref="GraphicsRenderPassObject" /> class.</summary>
-    /// <param name="renderPass">The render pass for which the object is being created.</param>
-    /// <exception cref="ArgumentNullException"><paramref name="renderPass" /> is <c>null</c>.</exception>
-    protected GraphicsRenderPassObject(GraphicsRenderPass renderPass)
+    /// <summary>Initializes a new instance of the <see cref="GraphicsPipelineObject" /> class.</summary>
+    /// <param name="pipeline">The pipeline for which the object is being created.</param>
+    /// <exception cref="ArgumentNullException"><paramref name="pipeline" /> is <c>null</c>.</exception>
+    protected GraphicsPipelineObject(GraphicsPipeline pipeline)
     {
-        ThrowIfNull(renderPass);
+        ThrowIfNull(pipeline);
 
-        _adapter = renderPass.Adapter;
-        _device = renderPass.Device;
-        _renderPass = renderPass;
-        _service = renderPass.Service;
+        _adapter = pipeline.Adapter;
+        _device = pipeline.Device;
+        _pipeline = pipeline;
+        _renderPass = pipeline.RenderPass;
+        _service = pipeline.Service;
     }
 
     /// <summary>Gets the underlying adapter for <see cref="Device" />.</summary>
@@ -32,10 +34,13 @@ public abstract class GraphicsRenderPassObject
     /// <summary>Gets the underlying device for <see cref="RenderPass" />.</summary>
     public GraphicsDevice Device => _device;
 
-    /// <summary>Gets the name for the render pass object.</summary>
+    /// <summary>Gets the name for the pipeline object.</summary>
     public abstract string Name { get; set; }
 
-    /// <summary>Gets the render pass for which the object was created.</summary>
+    /// <summary>Gets the pipeline for which the object was created.</summary>
+    public GraphicsPipeline Pipeline => _pipeline;
+
+    /// <summary>Gets the underlying render pass for <see cref="Pipeline" />.</summary>
     public GraphicsRenderPass RenderPass => _renderPass;
 
     /// <summary>Gets the underlying service for <see cref="Adapter" />.</summary>
