+++ sources/Graphics/Core/Graphics/GraphicsMemoryAllocator.DefaultMemoryAllocator.cs
@@ -11,6 +11,7 @@ using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using TerraFX.Collections;
+using TerraFX.Graphics.Advanced;
 using static TerraFX.Runtime.Configuration;
 using static TerraFX.Utilities.AssertionUtilities;
 using static TerraFX.Utilities.ExceptionUtilities;
