+++ sources/Graphics/Core/Graphics/GraphicsResource.cs
@@ -3,6 +3,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using TerraFX.Graphics.Advanced;
 
 namespace TerraFX.Graphics;
