+++ sources/Graphics/D3D12/Graphics/D3D12GraphicsRenderContext.cs
@@ -2,6 +2,7 @@
 
 using System;
 using System.Threading;
+using TerraFX.Graphics.Advanced;
 using TerraFX.Interop.DirectX;
 using TerraFX.Interop.Windows;
 using TerraFX.Numerics;
