+++ sources/Graphics/D3D12/Graphics/D3D12GraphicsRenderPass.cs
@@ -1,5 +1,6 @@
 // Copyright © Tanner Gooding and Contributors. Licensed under the MIT License (MIT). See License.md in the repository root for more information.
 
+using TerraFX.Graphics.Advanced;
 using TerraFX.Threading;
 using static TerraFX.Threading.VolatileState;
 using static TerraFX.Utilities.ExceptionUtilities;
