+++ sources/Graphics/D3D12/Graphics/D3D12GraphicsRenderTarget.cs
@@ -1,5 +1,6 @@
 // Copyright © Tanner Gooding and Contributors. Licensed under the MIT License (MIT). See License.md in the repository root for more information.
 
+using TerraFX.Graphics.Advanced;
 using TerraFX.Interop.DirectX;
 using TerraFX.Threading;
 using static TerraFX.Interop.DirectX.D3D12_RTV_DIMENSION;
