+++ sources/Graphics/Vulkan/Graphics/VulkanGraphicsBuffer.CreateInfo.cs
@@ -1,6 +1,7 @@
 // Copyright © Tanner Gooding and Contributors. Licensed under the MIT License (MIT). See License.md in the repository root for more information.
 
 using System.Runtime.InteropServices;
+using TerraFX.Graphics.Advanced;
 using TerraFX.Interop.Vulkan;
 
 namespace TerraFX.Graphics;
